python-chat-service/requirements.txt
fastapi==0.116.1
python-socketio==5.13.0
uvicorn==0.35.0
pg8000==1.31.5
python-jose==3.4.0
python-dotenv==1.1.1
